required changes: add docstrings to trial collection, move to separate file
